requirements.txt
fastapi
uvicorn[standard]
websockets
supabase
openai
python-dotenv
pydantic
asyncer
httpx
python-multipart